+++ CHANGELOG.md
@@ -9,7 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- badge with '+ additional titles' if there is more than one title present
+
 ### Changes
+
+- run local test browser in headed mode with higher resolution
 
 ### Deprecated